% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{common_val_scale}
\alias{common_val_scale}
\title{Get common value scale}
\usage{
common_val_scale(x, y)
}
\arguments{
\item{x}{First value scale.}

\item{y}{Second value scale.}
}
\value{
A character scalar.
}
\description{
Determines the greatest common \href{https://en.wikipedia.org/wiki/Level_of_measurement}{values scale} between two value scales \code{x} and \code{y} in the hierarchy
\emph{nominal} < \emph{ordinal} < \emph{interval} < \emph{ratio}.
}
\details{
Note that
\itemize{
\item the value scales \emph{interval} and \emph{ratio} are assumed to always be in \strong{ascending} order.
\item the value scale \emph{none} is treated specially in the sense that it is ignored when finding the common value scale between \emph{none} and any other value scale.
}
}
\examples{
qstnr::common_val_scale("interval", "nominal")
qstnr::common_val_scale("interval", "ratio")
qstnr::common_val_scale("ordinal_ascending", "ordinal_descending")

try(
  qstnr::common_val_scale("ordinal_descending", "ratio")
)
}
