% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{gen_qmd_qstnr}
\alias{gen_qmd_qstnr}
\title{Generate Quarto Markdown questionnaire}
\usage{
gen_qmd_qstnr(qstnr, survey_config, lang, path)
}
\arguments{
\item{qstnr}{Unnested questionnaire tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}}.}

\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey]{read_survey()}}.}

\item{lang}{Language in which to output the questionnaire. A character scalar.}

\item{path}{Path to write the generated questionnaire file to. A character scalar.}
}
\value{
\code{qstnr}, invisibly.
}
\description{
Generate Quarto Markdown questionnaire
}
