% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{gen_qstnr_docs}
\alias{gen_qstnr_docs}
\title{Generate questionnaire docs}
\usage{
gen_qstnr_docs(
  qstnr,
  survey_config,
  langs = NULL,
  add_item_ids = TRUE,
  output_dir,
  keep_qmd = TRUE,
  keep_docx = FALSE,
  create_g_docs = TRUE,
  g_drive_dir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{qstnr}{Unnested questionnaire tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}}.}

\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey_config]{read_survey_config()}}.}

\item{langs}{Languages in which to output the questionnaire. A subset of \code{survey_config$langs} or \code{NULL} for all languages present.}

\item{add_item_ids}{Whether or not to add questionnaire item identifiers next to the question texts.}

\item{output_dir}{Directory to write the generated questionnaire files to. A character scalar.}

\item{keep_qmd}{Whether to keep or delete the intermediate generated Quarto Markdown files.}

\item{keep_docx}{Whether to keep or delete the intermediate generated Microsoft Word files.}

\item{create_g_docs}{Whether or not to turn the generated files into Google Docs placed under the Google Drive path \code{g_drive_dir}.}

\item{g_drive_dir}{Google Drive path under which to place the generated Google Docs. By default, the file is created in the current user's "My Drive" root
folder. Ignored if \code{create_g_docs = FALSE}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{output_dir}, invisibly.
}
\description{
Generates a \link[=gen_qmd_qstnr]{Quarto Markdown representation} of the input for each specified language, renders it to Microsoft Word (DOCX) format and
optionally uploads it as a Google Doc to the specified \code{g_drive_dir}.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qmd_qstnr}()},
\code{\link{gen_qstnr}()},
\code{\link{htmlize_qstnr}()},
\code{\link{unnest_qstnr_vals}()},
\code{\link{unnest_qstnr}()}
}
\concept{qstnr_gen}
