% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{group_title}
\alias{group_title}
\title{Get survey group title}
\usage{
group_title(survey_config, group_id)
}
\arguments{
\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey_config]{read_survey_config()}}.}

\item{group_id}{Survey group identifier to retrieve the group title of. A character scalar.}
}
\value{
A list.
}
\description{
Extracts the multilanguage group title list for the specified group from the given survey config.
}
\seealso{
Other auxiliary functions:
\code{\link{as_int_vals}()},
\code{\link{common_val_scale}()},
\code{\link{val_set}()}
}
\concept{aux}
