% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{htmlize_qstnr}
\alias{htmlize_qstnr}
\title{Convert qstnr columns from Markdown to HTML}
\usage{
htmlize_qstnr(
  qstnr,
  cols = c("title", "description", "question"),
  cols_rm_p = "values",
  hardbreaks = FALSE,
  smart = TRUE,
  normalize = TRUE,
  extensions = c("table", "strikethrough", "tasklist")
)
}
\arguments{
\item{qstnr}{Semi-unnested questionnaire tibble as returned by \code{\link[=gen_qstnr]{gen_qstnr()}}.}

\item{cols}{Column names to convert to HTML. All columns must be of type character. Mustn't overlap with \code{cols_rm_p}.
\link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{cols_rm_p}{Columns names to convert to HTML \emph{without the outer \verb{<p>} tag}. All columns must be of type character. Mustn't overlap with \code{cols}.
\link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{hardbreaks}{Treat newlines as hard line breaks. If this option is specified, hard wrapping is disabled
regardless of the value given with \code{width}.}

\item{smart}{Use smart punctuation. See details.}

\item{normalize}{Consolidate adjacent text nodes.}

\item{extensions}{Enables Github extensions. Can be \code{TRUE} (all) \code{FALSE} (none) or a character
vector with a subset of available \link[commonmark]{extensions}.}
}
\value{
\code{qstnr} with values in \code{cols} converted to HTML text.
}
\description{
Recursively converts the specified columns from Markdown to HTML using \code{\link[commonmark:commonmark]{commonmark::markdown_html()}}. Columns to be converted must be of (nested) type
character.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qmd_qstnr}()},
\code{\link{gen_qstnr}()},
\code{\link{unnest_qstnr_vals}()},
\code{\link{unnest_qstnr}()}
}
\concept{qstnr_gen}
