% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{htmlize_survey_config}
\alias{htmlize_survey_config}
\title{Convert survey config from Markdown to HTML}
\usage{
htmlize_survey_config(
  survey_config,
  ...,
  rm_p = FALSE,
  hard_line_breaks = FALSE,
  smart_punctuation = TRUE,
  strip_footnotes = TRUE,
  normalize = TRUE,
  extensions = c("table", "strikethrough", "tasklist")
)
}
\arguments{
\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey_config]{read_survey_config()}}.}

\item{...}{\link[purrr:pluck]{Pluck} locations to convert to HTML. All (sub)keys targeted here must be of type character.
\link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{rm_p}{Whether or not to omit the outer \verb{<p>} tag when converting values to HTML.}

\item{hard_line_breaks}{Whether or not to interpret all newlines within a paragraph as hard line breaks instead of spaces. This setting conforms to \href{https://pandoc.org/MANUAL.html#extension-hard_line_breaks}{Pandoc's \code{hard_line_breaks} extension}.}

\item{smart_punctuation}{Whether or not to enable smart punctation conforming to \href{https://pandoc.org/MANUAL.html#extension-smart}{Pandoc's \code{smart} extension}
which converts straight quotes to curly quotes, \verb{---} to an em-dash (—), \verb{--} to an en-dash (–), and \code{...} to ellipses (…). It also replaces regular spaces
after certain abbreviations such as \code{Mr.} with \href{https://en.wikipedia.org/wiki/Non-breaking_space}{non-breaking spaces}.}

\item{strip_footnotes}{Whether or not to remove Markdown footnotes.}

\item{normalize}{Consolidate adjacent text nodes.}

\item{extensions}{Enables Github extensions. Can be \code{TRUE} (all) \code{FALSE} (none) or a character
vector with a subset of available \link[commonmark]{extensions}.}
}
\value{
\code{survey_config} with the specified keys recursively converted to HTML text.
}
\description{
Recursively converts the keys at the specified pluck locations of a survey configuration from Markdown to HTML using \code{\link[commonmark:commonmark]{commonmark::markdown_html()}}. Keys to
be converted must be of (nested) type character.
}
\seealso{
Other survey preparation functions:
\code{\link{read_survey_config}()}
}
\concept{survey_prep}
