% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{read_survey}
\alias{read_survey}
\title{Read in survey config}
\usage{
read_survey(path, validate = FALSE)
}
\arguments{
\item{path}{Path to the survey config TOML file. A character scalar.}

\item{validate}{Whether or not to validate all input TOML files, i.e. ensure that they adhere to their respective schemas.}
}
\value{
A list containing the combined survey configuration. Additionally,
}
\description{
Read in survey config
}
