% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{read_survey_config}
\alias{read_survey_config}
\title{Read in survey config}
\usage{
read_survey_config(path, validate = NULL)
}
\arguments{
\item{path}{Path to the survey config TOML file. A character scalar.}

\item{validate}{Whether or not to validate all input TOML files, i.e. ensure that they adhere to their respective schemas. Validation usually adds a
noticeable delay. If \code{NULL}, validation is performed only for files whose content has changed since the last successful validation.}
}
\value{
A list containing the combined survey configuration.
}
\description{
Assembles a raw TOML survey configuration into a combined survey configuration list.
}
\seealso{
Other survey preparation functions:
\code{\link{htmlize_survey_config}()}
}
\concept{survey_prep}
