% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{unnest_qstnr}
\alias{unnest_qstnr}
\title{Unnest questionnaire tibble}
\usage{
unnest_qstnr(qstnr)
}
\arguments{
\item{qstnr}{Nested questionnaire tibble as returned by \code{\link[=gen_qstnr]{gen_qstnr()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Unnest questionnaire tibble
}
