% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{unnest_qstnr_vals}
\alias{unnest_qstnr_vals}
\title{Unnest questionnaire tibble's item values}
\usage{
unnest_qstnr_vals(qstnr)
}
\arguments{
\item{qstnr}{Semi-unnested questionnaire tibble as returned by \code{\link[=gen_qstnr]{gen_qstnr()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Unnests a semi-unnested qstnr tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}} into one row per question \code{lang}, \code{id} and \code{value}.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qmd_qstnr}()},
\code{\link{gen_qstnr}()},
\code{\link{unnest_qstnr}()}
}
\concept{qstnr_gen}
