% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{val_set}
\alias{val_set}
\title{Get value set}
\usage{
val_set(id, survey_config = survey_config)
}
\arguments{
\item{id}{String that uniquely identifies the value set.}

\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey]{read_survey()}}.}
}
\value{
A list.
}
\description{
Returns the value set identified by the provided \code{id}. The prefix \code{"desc:"} can be used to reverse the order of the set's values.
}
\seealso{
Other auxiliary functions:
\code{\link{as_int_vals}()},
\code{\link{common_val_scale}()}
}
\concept{aux}
