% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{as_int_vals}
\alias{as_int_vals}
\title{Convert question values to codes}
\usage{
as_int_vals(
  var,
  qstnr = pal::pkg_config_val(key = "cur_qstnr", pkg = this_pkg)
)
}
\arguments{
\item{var}{Variable holding the question values. A character vector with an \code{id} and a \code{lang} attribute.}

\item{qstnr}{TODO}
}
\value{
An integer vector of the same length as \code{var}.
}
\description{
Converts character question values to their integer code counterparts. The latter are useful mainly because they're language-independent.
}
\seealso{
Other auxiliary functions:
\code{\link{common_val_scale}()},
\code{\link{group_title}()},
\code{\link{val_set}()}
}
\concept{aux}
