% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{gen_qmd_qstnr}
\alias{gen_qmd_qstnr}
\title{Generate Quarto Markdown questionnaire}
\usage{
gen_qmd_qstnr(qstnr, survey_config, lang, path, add_item_ids = TRUE)
}
\arguments{
\item{qstnr}{Unnested questionnaire tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}}.}

\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey_config]{read_survey_config()}}.}

\item{lang}{Language in which to output the questionnaire. A character scalar.}

\item{path}{Path to write the generated questionnaire file to. A character scalar.}

\item{add_item_ids}{Whether or not to add questionnaire item identifiers next to the question texts.}
}
\value{
\code{qstnr}, invisibly.
}
\description{
Generates a \link[=gen_qmd_qstnr]{Quarto Markdown representation} of the input.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qstnr}()},
\code{\link{gen_qstnr_docs}()},
\code{\link{htmlize_qstnr}()},
\code{\link{unnest_qstnr}()},
\code{\link{unnest_qstnr_vals}()}
}
\concept{qstnr_gen}
\keyword{internal}
