% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{gen_qstnr}
\alias{gen_qstnr}
\title{Generate questionnaire tibble}
\usage{
gen_qstnr(survey_config)
}
\arguments{
\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey]{read_survey()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Generate questionnaire tibble
}
