% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{htmlize_qstnr}
\alias{htmlize_qstnr}
\title{Convert qstnr columns from Markdown to HTML}
\usage{
htmlize_qstnr(
  qstnr,
  cols = c("title", "description", "question"),
  cols_rm_p = "values",
  hard_line_breaks = FALSE,
  smart_punctuation = TRUE,
  strip_footnotes = TRUE,
  normalize = TRUE,
  extensions = c("table", "strikethrough", "tasklist")
)
}
\arguments{
\item{qstnr}{Semi-unnested questionnaire tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}}.}

\item{cols}{Column names to convert to HTML. All columns must be of type character. Mustn't overlap with \code{cols_rm_p}.
\link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{cols_rm_p}{Columns names to convert to HTML \emph{without the outer \verb{<p>} tag}. All columns must be of type character. Mustn't overlap with \code{cols}.
\link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{hard_line_breaks}{Whether or not to interpret all newlines within a paragraph as hard line breaks instead of spaces. This setting conforms to \href{https://pandoc.org/MANUAL.html#extension-hard_line_breaks}{Pandoc's \code{hard_line_breaks} extension}.}

\item{smart_punctuation}{Whether or not to enable smart punctation conforming to \href{https://pandoc.org/MANUAL.html#extension-smart}{Pandoc's \code{smart} extension}
which converts straight quotes to curly quotes, \verb{---} to an em-dash (—), \verb{--} to an en-dash (–), and \code{...} to ellipses (…). It also replaces regular spaces
after certain abbreviations such as \code{Mr.} with \href{https://en.wikipedia.org/wiki/Non-breaking_space}{non-breaking spaces}.}

\item{strip_footnotes}{Whether or not to remove Markdown footnotes.}

\item{normalize}{Consolidate adjacent text nodes.}

\item{extensions}{Enables Github extensions. Can be \code{TRUE} (all) \code{FALSE} (none) or a character
vector with a subset of available \link[commonmark]{extensions}.}
}
\value{
\code{qstnr} with values in \code{cols} converted to HTML text.
}
\description{
Recursively converts the specified columns from Markdown to HTML using \code{\link[commonmark:commonmark]{commonmark::markdown_html()}}. Columns to be converted must be of (nested) type
character.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qmd_qstnr}()},
\code{\link{gen_qstnr_docs}()},
\code{\link{gen_qstnr}()},
\code{\link{unnest_qstnr_vals}()},
\code{\link{unnest_qstnr}()}
}
\concept{qstnr_gen}
