% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{htmlize_survey_config}
\alias{htmlize_survey_config}
\title{Convert survey config from Markdown to HTML}
\usage{
htmlize_survey_config(
  survey_config,
  ...,
  rm_p = FALSE,
  hardbreaks = FALSE,
  smart = TRUE,
  normalize = TRUE,
  extensions = c("table", "strikethrough", "tasklist")
)
}
\arguments{
\item{survey_config}{Survey configuration list as returned by \code{\link[=read_survey_config]{read_survey_config()}}.}

\item{...}{\link[purrr:pluck]{Pluck} locations to convert to HTML. All (sub)keys targeted here must be of type character.
\link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{rm_p}{Whether or not to omit the outer \verb{<p>} tag when converting values to HTML.}

\item{hardbreaks}{Treat newlines as hard line breaks. If this option is specified, hard wrapping is disabled
regardless of the value given with \code{width}.}

\item{smart}{Use smart punctuation. See details.}

\item{normalize}{Consolidate adjacent text nodes.}

\item{extensions}{Enables Github extensions. Can be \code{TRUE} (all) \code{FALSE} (none) or a character
vector with a subset of available \link[commonmark]{extensions}.}
}
\value{
\code{survey_config} with the specified keys recursively converted to HTML text.
}
\description{
Recursively converts the keys at the specified pluck locations of a survey configuration from Markdown to HTML using \code{\link[commonmark:commonmark]{commonmark::markdown_html()}}. Keys to
be converted must be of (nested) type character.
}
\seealso{
Other survey preparation functions:
\code{\link{read_survey_config}()}
}
\concept{survey_prep}
