% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr-package.R
\docType{package}
\name{qstnr-package}
\alias{qstnr}
\alias{qstnr-package}
\title{qstnr: Helps Creating and Maintaining Survey Questionnaires}
\description{
qstnr provides a framework to create and maintain survey questionnaires written in the TOML file format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qstnr.rpkg.dev}
  \item \url{https://gitlab.com/salim_b/r/pkgs/qstnr}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/qstnr/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
