% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr-package.R
\docType{package}
\name{qstnr-package}
\alias{qstnr}
\alias{qstnr-package}
\title{qstnr: Helps Creating and Maintaining Survey Questionnaires}
\description{
qstnr provides a framework to create and maintain survey questionnaires written in the TOML file format.
}
\section{Package configuration}{


Some of qstnr's functionality is controlled via package-specific global configuration which can either be set via \R \link{options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Maximal timespan to preserve the package's \href{https://rpkg.dev/pkgpins/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{qstnr.max_cache_lifespan} \tab \code{QSTNR_MAX_CACHE_LIFESPAN} \tab \code{"1 year"} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://qstnr.rpkg.dev}
  \item \url{https://gitlab.com/salim_b/r/pkgs/qstnr}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/qstnr/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
