% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstnr.gen.R
\name{unnest_qstnr_vals}
\alias{unnest_qstnr_vals}
\title{Unnest questionnaire tibble's item values}
\usage{
unnest_qstnr_vals(qstnr)
}
\arguments{
\item{qstnr}{Semi-unnested questionnaire tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Unnests a semi-unnested qstnr tibble as returned by \code{\link[=unnest_qstnr]{unnest_qstnr()}} into one row per question \code{lang}, \code{id} and \code{value}.

The columns \code{values}, \code{values_int} and \code{values_targets} will thereby become
\code{value}, \code{value_int} and \code{value_targets}. Additionally, a column \code{value_order} is added holding the intended
value order as defined in the questionnaire.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qmd_qstnr}()},
\code{\link{gen_qstnr_docs}()},
\code{\link{gen_qstnr}()},
\code{\link{htmlize_qstnr}()},
\code{\link{unnest_qstnr}()}
}
\concept{qstnr_gen}
